add range (#1052)
* add range
