+++ metadata/metadata.go
@@ -1,6 +1,8 @@
 package metadata
 
 import (
+	"context"
+	"fmt"
 	"strings"
 )
 
@@ -41,6 +43,16 @@ func (m Metadata) Set(key string, value string) {
 	m[k] = value
 }
 
+// Range iterate over element in metadata.
+func (m Metadata) Range(f func(k, v string) bool) {
+	for k, v := range m {
+		ret := f(k, v)
+		if ret == false {
+			break
+		}
+	}
+}
+
 // Clone returns a deep copy of Metadata
 func (m Metadata) Clone() Metadata {
 	md := Metadata{}
@@ -48,4 +60,54 @@ func (m Metadata) Clone() Metadata {
 		md[k] = v
 	}
 	return md
+}
+
+type serverMetadataKey struct{}
+
+// NewServerContext creates a new context with client md attached.
+func NewServerContext(ctx context.Context, md Metadata) context.Context {
+	return context.WithValue(ctx, serverMetadataKey{}, md)
+}
+
+// FromServerContext returns the server metadata in ctx if it exists.
+func FromServerContext(ctx context.Context) (Metadata, bool) {
+	md, ok := ctx.Value(serverMetadataKey{}).(Metadata)
+	return md, ok
+}
+
+type clientMetadataKey struct{}
+
+// NewClientContext creates a new context with client md attached.
+func NewClientContext(ctx context.Context, md Metadata) context.Context {
+	return context.WithValue(ctx, clientMetadataKey{}, md)
+}
+
+// FromClientContext returns the client metadata in ctx if it exists.
+func FromClientContext(ctx context.Context) (Metadata, bool) {
+	md, ok := ctx.Value(clientMetadataKey{}).(Metadata)
+	return md, ok
+}
+
+// AppendToClientContext returns a new context with the provided kv merged
+// with any existing metadata in the context.
+func AppendToClientContext(ctx context.Context, kv ...string) context.Context {
+	if len(kv)%2 == 1 {
+		panic(fmt.Sprintf("metadata: AppendToOutgoingContext got an odd number of input pairs for metadata: %d", len(kv)))
+	}
+	md, _ := FromClientContext(ctx)
+	md = md.Clone()
+	for i := 0; i < len(kv); i += 2 {
+		md.Set(kv[i], kv[i+1])
+	}
+	return NewClientContext(ctx, md)
+}
+
+// MergeToClientContext merge new metadata into ctx.
+func MergeToClientContext(ctx context.Context, cmd Metadata) context.Context {
+	md, _ := FromClientContext(ctx)
+	md = md.Clone()
+	for k, v := range cmd {
+		md[k] = v
+	}
+	return NewClientContext(ctx, md)
 }
